requirements.txt
pillow==11.0.0
pytesseract==0.3.13
cairosvg==2.7.1
